Search Form state and functions moved to CharacterList.js and basic search functionality working. Still needs improving.
