+++ src/components/CharacterList.js
@@ -18,27 +18,71 @@ export default function CharacterList() {
       })
   }, []);
 
+  // Search Form State and Functions
+  const [name, setName] = useState("")
+  const [results, setResults] = useState("")
+
+  const handleChange = event => {
+    setName(event.target.value)
+  }
+
+  const handleSubmit = event => {
+    console.log("Submit Pressed!")  
+    event.preventDefault();   
+    // 
+    const searchResults = data.filter(char => {
+      return char.name.includes(name);
+    })
+    setResults(searchResults);
+    console.log("Searched Name", searchResults)
+    console.log("Results", results)
+    // setName("");
+  }
+
   return (
     <div className="cards-container">
       <div>
-        <SearchForm List={data} />
-      </div><br></br>
+        <SearchForm 
+        name={name} 
+        results={results} 
+        handleChange={handleChange}
+        handleSubmit={handleSubmit}
+         />
+      </div>
+      {/* adding space between search form and cards */}
+      <br></br> 
       <br></br>
-      <br></br>      
+      <br></br>   
+
       <section className="character-list grid-view">
-        {data.map(char => {
-          return (
-            <CharacterCard 
-              key={char.id}
-              image={char.image}
-              name={char.name}
-              species={char.species}
-              gender={char.gender}
-              status={char.status}
-            />
-          )        
-        })
-      }
+          {results !== "" ? 
+            results.map(char => {
+              return (
+                <CharacterCard 
+                  key={char.id}
+                  image={char.image}
+                  name={char.name}
+                  species={char.species}
+                  gender={char.gender}
+                  status={char.status}
+                />
+              )
+            }) 
+            // else
+            : 
+            data.map(char => {
+              return (
+                <CharacterCard 
+                  key={char.id}
+                  image={char.image}
+                  name={char.name}
+                  species={char.species}
+                  gender={char.gender}
+                  status={char.status}
+                />
+              )        
+            })
+          }
       </section>
     </div>
   );
