+++ src/components/SearchForm.js
@@ -1,42 +1,23 @@
-import React, { useState } from "react";
+import React from "react";
 import { StyledForm } from '../Styles.js';
 import CharacterCard from "./CharacterCard";
 
 
-export default function SearchForm(props) {
-  const [name, setName] = useState("")
-  const [results, setResults] = useState([props.list])
-
-  const handleChange = event => {
-    setName(event.target.value)
-  }
-
-  const handleSubmit = event => {
-    console.log("Submit Pressed!", props)
-    console.log("Results", results)
-    event.preventDefault();
-    setName("");
-    const searchedName = results.filter(char => {
-      // return char.name === name;
-      console.log(char); // empty...
-    })
-    console.log(searchedName);
-  }
+export default function SearchForm(props) {  
  
   return (
-    <section className="search-form">      
-      <StyledForm onSubmit={handleSubmit}>
+    <section className="search-form">    
+      <StyledForm onSubmit={props.handleSubmit}>
         <input 
           id="name"
-          value={name}
+          value={props.name}
           name="name"
           type="text"
           placeholder="Search for character by Name"
-          onChange={handleChange}
+          onChange={props.handleChange}
         />
         <button type="submit">Search</button>
       </StyledForm>
-      {name}
-    </section>
-  );
+    </section> 
+  )  
 }
