+++ src/index.css
@@ -5,9 +5,9 @@ body {
 }
 
 main {
-  width: 85vw;
+  width: 95vw;
   min-height: 80vh;
-  max-width: 1024px;
+  max-width: 1424px;
   margin: 35px auto;
   background: #fff;
   padding: 15px;
